Move FileUtils code from Driver to FileUtils and use java.nio methods.
